some changes
I changed importconfig to config and changed the way imports work to avoid using from blah import *. So now we just import config and use config.variablename. This let me put the whole disocrd handler in a function and call it with multiprocessing.

Kill log is functional, made some additionals to db table fields for wanted, and bounty info